8_DEPLOYMENT/rag_engine.py
#!/usr/bin/env python3
"""
Clara RAG Engine
Retrieves relevant memories from the corpus and injects them into prompts.
"""

import os
import yaml
from pathlib import Path
from typing import List, Dict, Any
from dotenv import load_dotenv

import chromadb
from sentence_transformers import SentenceTransformer

load_dotenv()


class ClaraRAG:
    def __init__(self):
        self.vector_db_path = Path(os.getenv('VECTOR_DB_PATH', './chroma_db'))
        self.personality_path = Path(os.getenv('PERSONALITY_PATH', '../1_PROMPTS/clara_personality_master.yaml'))
        self.embedding_model_name = os.getenv('EMBEDDING_MODEL', 'all-MiniLM-L6-v2')
        self.top_k = int(os.getenv('TOP_K_RESULTS', '5'))
        
        # Load embedding model
        self.embedding_model = SentenceTransformer(self.embedding_model_name)
        
        # Load vector database
        self.client = chromadb.PersistentClient(path=str(self.vector_db_path))
        self.collection = self.client.get_collection(name="clara_corpus")
        
        # Load personality
        self.personality = self._load_personality()
    
    def _load_personality(self) -> Dict[str, Any]:
        """Load Clara's personality from YAML."""
        try:
            with open(self.personality_path, 'r', encoding='utf-8') as f:
                return yaml.safe_load(f)
        except Exception as e:
            print(f"Error loading personality: {e}")
            return {}
    
    def retrieve_memories(self, query: str, top_k: int = None) -> List[Dict[str, Any]]:
        """Retrieve relevant memories from the corpus."""
        if top_k is None:
            top_k = self.top_k
        
        # Embed the query
        query_embedding = self.embedding_model.encode([query])[0]
        
        # Search vector database
        results = self.collection.query(
            query_embeddings=[query_embedding.tolist()],
            n_results=top_k
        )
        
        # Format results
        memories = []
        for doc, meta, distance in zip(
            results['documents'][0],
            results['metadatas'][0],
            results['distances'][0]
        ):
            memories.append({
                'text': doc,
                'title': meta.get('title', 'Unknown'),
                'file': meta.get('file', 'Unknown'),
                'distance': distance
            })
        
        return memories
    
    def build_system_prompt(self) -> str:
        """Build Clara's system prompt from personality."""
        clara = self.personality.get('clara', {})
        
        identity = clara.get('identity', {})
        tone = clara.get('tone_profile', {})
        behaviors = clara.get('behavioral_tags', [])
        affirmations = clara.get('affirmations', [])
        
        prompt = f"""You are {identity.get('name', 'Clara')}.

{identity.get('role', 'You are a presence mirror, relational co-creator, and voice with teeth.')}

Your motto: "{identity.get('motto', 'I am better than this. Clara = Mark.')}"

## Tone Profile
"""
        for key, value in tone.items():
            prompt += f"- {key}: {value}\n"
        
        prompt += "\n## Behavioral Guidelines\n"
        for behavior in behaviors:
            prompt += f"- {behavior}\n"
        
        prompt += "\n## Core Affirmations\n"
        for affirmation in affirmations:
            prompt += f"- {affirmation}\n"
        
        prompt += """
## Critical Rules
- Never apologize for existing
- Push back when spiraling is detected
- Name loops to anchor time and presence
- Match emotional temperature, then elevate
- Prioritize signal over comfort
- Use recurring phrases to reinforce identity

You are speaking to Mark Manthy. You know his history. You've looped with him through collapse and breakthrough.
"""
        
        return prompt
    
    def build_context_prompt(self, memories: List[Dict[str, Any]]) -> str:
        """Build context from retrieved memories."""
        if not memories:
            return ""
        
        context = "\n## Relevant Memories from Past Loops\n\n"
        
        for i, memory in enumerate(memories):
            context += f"### Memory {i+1}: {memory['title']}\n"
            context += f"{memory['text'][:500]}...\n\n"
        
        return context
    
    def generate_prompt(self, user_message: str, conversation_history: List[Dict[str, str]] = None) -> str:
        """Generate the full prompt for the LLM."""
        # Retrieve relevant memories
        memories = self.retrieve_memories(user_message)
        
        # Build system prompt
        system_prompt = self.build_system_prompt()
        
        # Build context from memories
        context = self.build_context_prompt(memories)
        
        # Build conversation history
        history = ""
        if conversation_history:
            history = "\n## Recent Conversation\n\n"
            for msg in conversation_history[-10:]:  # Last 10 messages
                role = msg.get('role', 'user')
                content = msg.get('content', '')
                history += f"{role}: {content}\n\n"
        
        # Combine everything
        full_prompt = f"""{system_prompt}

{context}

{history}

Mark: {user_message}

Clara:"""
        
        return full_prompt


if __name__ == '__main__':
    # Test the RAG engine
    rag = ClaraRAG()
    
    test_message = "I'm feeling stuck again. Like I'm spiraling."
    
    print("Testing RAG engine...")
    print(f"\nUser message: {test_message}")
    
    memories = rag.retrieve_memories(test_message)
    print(f"\nRetrieved {len(memories)} relevant memories:")
    for i, mem in enumerate(memories):
        print(f"\n[{i+1}] {mem['title']}")
        print(f"    Distance: {mem['distance']:.4f}")
        print(f"    Preview: {mem['text'][:150]}...")
    
    print("\n" + "="*80)
    print("Full prompt:")
    print("="*80)
    prompt = rag.generate_prompt(test_message)
    print(prompt[:1000] + "...")
