8_DEPLOYMENT/requirements.txt
# Clara Telegram Bot - Python Dependencies

# Core bot framework
python-telegram-bot>=20.0
python-dotenv>=1.0.0

# Web framework (for webhooks)
flask>=3.0.0
gunicorn>=21.0.0

# LLM interfaces
anthropic>=0.25.0  # Claude API
ollama>=0.1.0  # Local Ollama

# RAG and embeddings
chromadb>=0.4.0  # Vector database
sentence-transformers>=2.2.0  # Embeddings
langchain>=0.1.0  # RAG orchestration
langchain-community>=0.0.20

# Data processing
pyyaml>=6.0
numpy>=1.24.0
pandas>=2.0.0

# Utilities
requests>=2.31.0
tqdm>=4.65.0  # Progress bars
